+++ src/masterchat.ts
@@ -1,5 +1,6 @@
 import axios, { AxiosError, AxiosInstance, AxiosRequestConfig } from "axios";
 import { EventEmitter } from "events";
+import { AsyncIterator } from "iterator-helpers-polyfill";
 import { buildMeta } from "./api";
 import { buildAuthHeaders } from "./auth";
 import { parseAction } from "./chat";
@@ -585,12 +586,10 @@ export class Masterchat extends EventEmitter {
   /**
    * AsyncIterator API
    */
-  public async *iter(options?: IterateChatOptions) {
-    for await (const { actions } of this.iterate(options)) {
-      for (const action of actions) {
-        yield action;
-      }
-    }
+  public iter(options?: IterateChatOptions): AsyncIterator<Action> {
+    return AsyncIterator.from<ChatResponse>(
+      this.iterate(options)
+    ).flatMap<Action>((action) => action.actions);
   }
 
   /**